clang

+++ linbox/linbox/algorithms/rational-solver2.h
@@ -208,7 +208,7 @@ namespace LinBox
 	{
 		if (n < 1) return 0;
 		double* IM = new double[n * n];
-		memcpy ((void*)IM, (const void*)M, sizeof(double)*n*n);
+		memcpy ((void*)IM, (const void*)M, sizeof(double)*(size_t)(n*n));
 		int ret;
 		//compute the inverse by flops
 		ret = cblas_dgeinv (IM, n);
@@ -243,7 +243,7 @@ namespace LinBox
 		//double log2 = log (2.0);
 		double log2 = M_LN2;
 		// r = b
-		memcpy ((void*) r, (const void*) b, sizeof(double)*n);
+		memcpy ((void*) r, (const void*) b, sizeof(double)*(size_t)n);
 
 		do  {
 			cblas_dapply (n, n, IM, r, x);
